move api calls out of Layout
